+++ packages/webxr-input/src/index.ts
@@ -15,3 +15,5 @@ export * from "./capabilities.js";
 export * from "./source.js";
 export * from "./provider.js";
 export * from "./pointer.js";
+export * from "./velocity.js";
+export * from "./contract-cases.js";